Move ANTI_FORGERY_KEY to object
