+++ app/controllers/Login.scala
@@ -1,6 +1,7 @@
 package controllers
 
-import actions.CSRFAction
+import csrf.CSRFAction
+import csrf.CSRF.ANTI_FORGERY_KEY
 import play.api.i18n.Messages
 import play.api.i18n.Messages.Implicits._
 import play.api.libs.json.Json
@@ -18,7 +19,6 @@ trait AuthActions extends Actions {
 }
 
 object Login extends Controller with AuthActions {
-  val ANTI_FORGERY_KEY = current.configuration.getString("identity-admin.antiForgeryKey").get
   val clientId = current.configuration.getString("identity-admin.google.clientId").get
   val clientSecret = current.configuration.getString("identity-admin.google.clientSecret").get
   val redirectUrl = current.configuration.getString("identity-admin.google.authorisationCallback").get
