+++ app/csrf/CSRFAction.scala
@@ -1,4 +1,4 @@
-package actions
+package csrf
 
 import controllers.routes
 import play.api.i18n.Messages
@@ -13,14 +13,17 @@ import scala.concurrent.Future
 class CSRFRequest[A](val csrfToken: String, request: Request[A]) extends WrappedRequest[A](request)
 
 object CSRFAction extends ActionRefiner[Request, CSRFRequest] {
-  private val ANTI_FORGERY_KEY = current.configuration.getString("identity-admin.antiForgeryKey").get
   private val loginCall = routes.Login.login()
 
   override protected def refine[A](request: Request[A]): Future[Either[Result, CSRFRequest[A]]] = Future {
-    val csrfToken = request.session.get(ANTI_FORGERY_KEY)
+    val csrfToken = request.session.get(CSRF.ANTI_FORGERY_KEY)
     csrfToken match {
       case Some(token) => Right(new CSRFRequest[A](token, request))
       case None => Left(Redirect(loginCall).flashing("error" -> Messages("login.missingAntiForgeryToken")))
     }
   }
+}
+
+object CSRF {
+  val ANTI_FORGERY_KEY = "antiForgeryToken"
 }
