+++ conf/CODE.conf
@@ -4,5 +4,4 @@ identity-admin {
   google {
     authorisationCallback = "https://useradmin.code.dev-gutools.co.uk/oauth2callback"
   }
-  antiForgeryKey = "antiForgeryToken"
 }
