+++ conf/DEV.conf
@@ -4,5 +4,4 @@ identity-admin {
   google {
     authorisationCallback = "https://useradmin.thegulocal.com/oauth2callback"
   }
-  antiForgeryKey = "antiForgeryToken"
 }
